Change read_sd_file semantics to return size read.

+++ fusee/fusee-primary/src/sd_utils.c
@@ -1,6 +1,6 @@
 #include "sd_utils.h"
 
-int read_sd_file(void *dst, size_t dst_size, const char *filename) {
+size_t read_sd_file(void *dst, size_t dst_size, const char *filename) {
     /* TODO: Implement this function. */
     (void)(dst);
     (void)(dst_size);
